Handle parsing of reports with no .Results list
A trivy fs scan may have no .Results list if there are no manifests for
it to process. This might happen with an empty directory, for example.

This commit allows the report_generator.parse_results function to return
None if there is no Results key in the report data. print_issue will
test that the value returned by parse_results is not None before
proceeding.

Fixes #1.